change int to long

+++ src/lesson9/lab9/prob2/PrimeStream.java
@@ -6,7 +6,7 @@ import java.util.stream.Stream;
 public class PrimeStream {
 	
 
-	void printFirstNPrimes(int n){
+	void printFirstNPrimes(long n){
 		Stream.iterate(BigInteger.valueOf(2), number -> number.nextProbablePrime()).limit(n).forEach(i -> System.out.println(i));
 	}
